requirements.txt
Pillow==9.3.0
Jinja2==3.1.2